Removed fmt dependency, added some math types and some utilities

+++ src/Rng.cpp
@@ -1,4 +1,5 @@
 #include "Hq/Rng.h"
+#include "Hq/Types.h"
 
 #include <ctime>
 #include <random>
@@ -6,7 +7,7 @@
 namespace hq
 {
 static auto                                  seed = time(nullptr);
-static std::mt19937                          gRng(seed);
+static std::mt19937                          gRng(static_cast<uint>(seed));
 static std::uniform_real_distribution<float> gDistribution;
 
 float rand01()
